pyproject.toml
[tool.poetry]
name = "currency-project"
version = "0.1.0"
description = ""
authors = ["Dontgoingforyou <mishael000@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
requests = "^2.32.3"
django-background-tasks = "^1.2.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
